chore(assistant builder): cleanup by moving some things around
