+++ front/components/assistant_builder/AssistantBuilder.tsx
@@ -3,11 +3,9 @@ import "react-image-crop/dist/ReactCrop.css";
 import {
   BuilderLayout,
   Button,
-  ChatBubbleBottomCenterTextIcon,
   CircleIcon,
   SquareIcon,
   Tab,
-  TemplateIcon,
   TriangleIcon,
 } from "@dust-tt/sparkle";
 import type {
@@ -30,15 +28,11 @@ import type * as t from "io-ts";
 import { useRouter } from "next/router";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import React from "react";
-import ReactMarkdown from "react-markdown";
 import { useSWRConfig } from "swr";
 
 import { SharingButton } from "@app/components/assistant/Sharing";
-import {
-  TryAssistant,
-  usePreviewAssistant,
-} from "@app/components/assistant/TryAssistant";
 import ActionScreen from "@app/components/assistant_builder/ActionScreen";
+import AssistantBuilderPreviewDrawer from "@app/components/assistant_builder/AssistantBuilderPreviewDrawer";
 import { InstructionScreen } from "@app/components/assistant_builder/InstructionScreen";
 import NamingScreen, {
   removeLeadingAt,
@@ -61,7 +55,6 @@ import { subNavigationBuild } from "@app/components/sparkle/navigation";
 import { SendNotificationsContext } from "@app/components/sparkle/Notification";
 import { isUpgraded } from "@app/lib/plans/plan_codes";
 import { useSlackChannelsLinkedWithAgent } from "@app/lib/swr";
-import { classNames } from "@app/lib/utils";
 import type { FetchAssistantTemplateResponse } from "@app/pages/api/w/[wId]/assistant/builder/templates/[tId]";
 
 type SlackChannel = { slackChannelId: string; slackChannelName: string };
@@ -419,38 +412,6 @@ export default function AssistantBuilder({
     ? `Edit @${builderState.handle}`
     : "New Assistant";
 
-  const {
-    shouldAnimate: shouldAnimatePreviewDrawer,
-    draftAssistant,
-    isFading,
-  } = usePreviewAssistant({ owner, builderState });
-
-  const [previewDrawerCurrentTab, setPreviewDrawerCurrentTab] = useState<
-    "Preview" | "Template"
-  >(template ? "Template" : "Preview");
-
-  const previewDrawerTabs = useMemo(
-    () => [
-      {
-        label: "Preview",
-        current: previewDrawerCurrentTab === "Preview",
-        onClick: () => {
-          setPreviewDrawerCurrentTab("Preview");
-        },
-        icon: ChatBubbleBottomCenterTextIcon,
-      },
-      {
-        label: "Template",
-        current: previewDrawerCurrentTab === "Template",
-        onClick: () => {
-          setPreviewDrawerCurrentTab("Template");
-        },
-        icon: TemplateIcon,
-      },
-    ],
-    [previewDrawerCurrentTab]
-  );
-
   return (
     <>
       <AppLayout
@@ -555,78 +516,12 @@ export default function AssistantBuilder({
             </div>
           }
           rightPanel={
-            <div className="h-full pb-5">
-              {template ? (
-                <Tab
-                  tabs={previewDrawerTabs}
-                  variant="default"
-                  className="hidden lg:flex"
-                />
-              ) : null}
-              {previewDrawerCurrentTab === "Preview" ? (
-                <div
-                  className={classNames(
-                    "flex h-full w-full overflow-hidden rounded-xl border border-structure-200 bg-structure-50 transition-all",
-                    shouldAnimatePreviewDrawer &&
-                      previewDrawerOpenedAt != null &&
-                      // Only animate the reload if the drawer has been open for at least 1 second.
-                      // This is to prevent the animation from triggering right after the drawer is opened.
-                      Date.now() - previewDrawerOpenedAt > 1000
-                      ? "animate-reload"
-                      : ""
-                  )}
-                >
-                  <TryAssistant
-                    owner={owner}
-                    assistant={draftAssistant}
-                    conversationFading={isFading}
-                  />
-                </div>
-              ) : (
-                <ReactMarkdown
-                  components={{
-                    h1: ({ children }) => (
-                      <h1 className="pb-2 pt-4 text-5xl font-semibold text-element-900">
-                        {children}
-                      </h1>
-                    ),
-                    h2: ({ children }) => (
-                      <h2 className="pb-2 pt-4 text-4xl font-semibold text-element-900">
-                        {children}
-                      </h2>
-                    ),
-                    h3: ({ children }) => (
-                      <h3 className="pb-2 pt-4 text-2xl font-semibold text-element-900">
-                        {children}
-                      </h3>
-                    ),
-                    strong: ({ children }) => (
-                      <strong className="font-semibold text-element-900">
-                        {children}
-                      </strong>
-                    ),
-                    ul: ({ children }) => (
-                      <ul className="list-disc py-2 pl-8 text-element-800 first:pt-0 last:pb-0">
-                        {children}
-                      </ul>
-                    ),
-                    ol: ({ children }) => (
-                      <ol className="list-decimal py-3 pl-8 text-element-800 first:pt-0 last:pb-0">
-                        {children}
-                      </ol>
-                    ),
-                    li: ({ children }) => (
-                      <li className="py-2 text-element-800 first:pt-0 last:pb-0">
-                        {children}
-                      </li>
-                    ),
-                  }}
-                  className="pr-8 pt-4"
-                >
-                  {template?.helpInstructions ?? ""}
-                </ReactMarkdown>
-              )}
-            </div>
+            <AssistantBuilderPreviewDrawer
+              template={template}
+              owner={owner}
+              previewDrawerOpenedAt={previewDrawerOpenedAt}
+              builderState={builderState}
+            />
           }
           isRightPanelOpen={previewDrawerOpenedAt !== null}
           toggleRightPanel={
